nlpfinal/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
streamlit>=1.33,<2
torch>=2.2,<3
transformers>=4.39,<5
pillow>=10,<12
sentence-transformers>=2.6,<4
opencv-python-headless>=4.9,<5
numpy>=1.26,<3
spacy>=3.7,<4
